requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
gunicorn>=22.0
polib>=1.2
django-filter>=24.0
python-dotenv>=1.0
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3,<6.0
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
